+++ junos/func_common.go
@@ -218,6 +218,25 @@ func uniqueListString(s []string) []string {
 	return r
 }
 
+type sortStringsLength []string
+
+func (s sortStringsLength) Len() int {
+	return len(s)
+}
+func (s sortStringsLength) Swap(i, j int) {
+	s[i], s[j] = s[j], s[i]
+}
+func (s sortStringsLength) Less(i, j int) bool {
+	if len(s[i]) < len(s[j]) {
+		return true
+	}
+	if len(s[j]) < len(s[i]) {
+		return false
+	}
+
+	return s[i] < s[j]
+}
+
 func checkStringHasPrefixInList(s string, list []string) bool {
 	for _, item := range list {
 		if strings.HasPrefix(s, item) {